Add support for storage area onChanged event listeners
Aside from the onChanged events on the storage namespace, each storage
area (sync, local and managed) have their own onChanged event. This
event is tied to the area and as such, it has a different callback
function signature that only provides an object mapping with the
relevant StorageChange(s).

References:
https://developer.chrome.com/apps/storage#type-StorageArea
https://chromium.googlesource.com/chromium/src/+/master/extensions/common/api/storage.json#152